Add color by stranded rna-seq demo
