+++ products/jbrowse-web/src/tests/__image_snapshots__/alignments-color-by-test-tsx-color-by-tag-1-snap.png
