+++ products/jbrowse-web/src/tests/__image_snapshots__/alignments-filters-test-tsx-filter-by-hp-tag-bam-1-snap.png
