+++ products/jbrowse-web/src/tests/__image_snapshots__/alignments-filters-test-tsx-filter-by-hp-tag-cram-1-snap.png
