+++ products/jbrowse-web/src/tests/__image_snapshots__/alignments-soft-clip-test-tsx-opens-the-track-menu-and-enables-soft-clipping-1-snap.png
