+++ products/jbrowse-web/src/tests/__image_snapshots__/alignments-sort-test-tsx-selects-a-sort-sort-by-base-pair-1-snap.png
